0x59fCDc21cadb367f57E8bC0915D94d2bFBdcE677
infinity token is a token established through the binance smart chain . its purpose is to make money decentralized and more convenient for you. for this purpose, we will activate a poker game that we will develop and get you will be able to make high profits. You will be given a daily limit, but you you can also send #1nfi token to the poker game and play the game